\name{givemeall-package}
\alias{givemeall-package}
\alias{givemeall}
\docType{package}
\title{Give-Me-All Auxiliary package}
\description{Auxiliary package which can be used to install all MO packages at once.}
\details{
\tabular{ll}{
Package: \tab givemeall\cr
Type: \tab Package\cr
Version: \tab 0.02\cr
Date: \tab 2017-03-14\cr
License: \tab BSD_2_clause + file LICENSE\cr
LazyLoad: \tab no\cr
}
}
\author{
Jan Philipp Dietrich

Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
